Extracting code from main.js to several files exposing their intent
