+++ js/windows.js
@@ -1,0 +1,15 @@
+// Keep a global reference of the window object, if you don't, the window will
+// be closed automatically when the JavaScript object is garbage collected.
+let mainWindow = null;
+let waiverWindow = null;
+let prefWindow = null;
+let tray = null;
+let contextMenu = null;
+
+module.exports = {
+    mainWindow,
+    waiverWindow,
+    prefWindow,
+    tray,
+    contextMenu
+};
